+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Fixed issue with `setFillMode` when using cross-domain URLs in iframe @rafa8626
 * Added new `proxyType` and `streamDelimiter` variables to Flash video to support RTMPS compatibility @rafa8626
+* Fixed issue with `Dailymotion` renderer when trying to use `muted` property @rafa8626
 
 *4.2.6 (2017/09/19)*